Refine distill reader interface
